requirements.txt
astroid==2.9.3
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.4
firebase-admin==5.2.0
Flask==2.0.3
google-api-core==2.2.2
google-api-python-client==2.29.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.53.0
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.11.3
isort==5.10.1
itsdangerous==2.1.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.0
mccabe==0.6.1
msgpack==1.0.3
oauthlib==3.1.1
packaging==21.3
platformdirs==2.5.1
proto-plus==1.20.3
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.12.2
pyparsing==3.0.5
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
toml==0.10.2
typed-ast==1.5.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
